Added a .syntastic_cpp_config file.
[Syntastic][1] is a vim plugin for syntax checking. By default,
syntastic gets a bit confused when syntax checking C++; it doesn't know
exactly where to find all the included files. To make syntastic usable,
we have to add a .syntastic_cpp_config file with the locations of all
the header files we include. The header include paths are taken from the
root `CMakeLists.txt` file.

[1]: https://github.com/vim-syntastic/syntastic