Removed example directory.
When the project was just getting started, I included an `example`
directory as an example of how to organize files, include third party
libraries, write tests, write benchmakrs, and so on. Now that the
project is far along, the directory doesn't serve much purpose. Plus,
it's probably easy to confuse with the `examples` directory. So, I just
deleted it.