Fixed stdlib oddity + fixed scratch merge.
Go to [wandbox](http://melpon.org/wandbox/) and enter in the following
code. Try it out with different compiler versions.

```cpp

int main() {
    std::set<int> xs;
    std::set<int> ys;
    ys.insert(std::make_move_iterator(xs.begin()),
              std::make_move_iterator(xs.end()));
}
```

It will compile for a lot of versions, but for some (e.g. clang 3.8.1)
it won't. It's a bit unclear what exactly the problem is but it seems
like `xs.begin()` and `xs.end()` are const iterators and making
move iterators out of them is a no-no. Changing `xs` to type `vector`
will make everything build.

Fluent had code that looked like this. We were streaming relational
algebra expressions into a buffer and then moving the buffer into a set.
This was building for me but not for @cw75. This commit changes things
to vectors so that they build for both of us.

I also refactored the `Scratch::Merge` implementation to use the
`ra_util.h` functions. I added some unit tests for `ra_util.h` too!