Revised serialization code.
Previously, serialization between channels was done with global ToString
and FromString methods. Relying on a global function is not very
flexible and also pretty confusing.

This commit introduces a MockPickler struct template which can be used
to pickle and unpickle C++ objects. It has the exact same functionality
of the global ToString and FromString functions from before, but now
it's not some global thing.

All the code that does serialization (e.g. Channel, FluentExecutor) is
now parameterized on a Pickler.