Added script to launch chat service.
Starting the chat service is pretty annoying. First, you have to reset
the database. Then you have to start the server and a bunch of clients.
Each command is super long and easy to ignore. It's a hassle.

I wrote a script which automatically creates a tmux window and launches
a full chat service in it:

        ./scripts/launch_chat_service.sh